Added ESLint and Prettier
Made ESLint use Prettier and also formatted the whole project with
prettier
