+++ .eslintrc.js
@@ -1,29 +1,32 @@
 module.exports = {
-  env: {
-    browser: true,
-    es6: true,
-  },
-  extends: [
-    'plugin:react/recommended',
-    'airbnb',
-  ],
-  globals: {
-    Atomics: 'readonly',
-    SharedArrayBuffer: 'readonly',
-  },
-  parserOptions: {
-    ecmaFeatures: {
-      jsx: true,
+    env: {
+        browser: true,
+        es6: true,
     },
-    ecmaVersion: 2018,
-    sourceType: 'module',
-  },
-  plugins: [
-    'react',
-  ],
-  rules: {
-    "no-shadow": "warn",
-    "no-param-reassign": "warn",
-    "no-restricted-globals": "warn"
-  },
+    extends: [
+        'plugin:react/recommended',
+        'airbnb',
+        'plugin:prettier/recommended',
+        'prettier/react'
+    ],
+    globals: {
+        Atomics: 'readonly',
+        SharedArrayBuffer: 'readonly',
+    },
+    parserOptions: {
+        ecmaFeatures: {
+            jsx: true,
+        },
+        ecmaVersion: 2018,
+        sourceType: 'module',
+    },
+    plugins: [
+        'react',
+    ],
+    rules: {
+        "no-shadow": "warn",
+        "no-param-reassign": "warn",
+        "no-restricted-globals": "warn",
+        curly: [2, "all"]
+    },
 };
