+++ app/components/AppTitle.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { View, Text } from 'react-native';
-import { bodyStyle } from '../Styles/style';
+import { bodyStyle } from '../styles/style';
 import { MORTGAGE_TITLE } from '../constants/constants';
 
 const ApplicationTitle = () => (
-  <View style={bodyStyle.TitleContainer}>
-    <Text style={bodyStyle.Title}>{MORTGAGE_TITLE}</Text>
-  </View>
+	<View style={bodyStyle.TitleContainer}>
+		<Text style={bodyStyle.Title}>{MORTGAGE_TITLE}</Text>
+	</View>
 );
 
 export default ApplicationTitle;
