+++ app/components/FormLine.jsx
@@ -2,40 +2,40 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Text, View } from 'react-native';
 import NumberTextInput from './NumberTextInput';
-import { inputStyle } from '../Styles/style';
+import { inputStyle } from '../styles/style';
 
 const FormLine = ({ label, value, valueUpdateCB }) => {
-  // console.log('VALUE UPDATE CB IN FORM LINE ' + valueUpdateCB);
-  // console.log('VALUE UPDATE CB IN FORM LINE ' + typeof valueUpdateCB);
-  const removeNonNumericValues = (label, value) => {
-    if (!value) {
-      value = 0;
-    } else {
-      const regexNotNumberOrDot = /[^0-9.]/g;
-      const regexLeadingNumber = /^0+/;
+	// console.log('VALUE UPDATE CB IN FORM LINE ' + valueUpdateCB);
+	// console.log('VALUE UPDATE CB IN FORM LINE ' + typeof valueUpdateCB);
+	const removeNonNumericValues = (label, value) => {
+		if (!value) {
+			value = 0;
+		} else {
+			const regexNotNumberOrDot = /[^0-9.]/g;
+			const regexLeadingNumber = /^0+/;
 
-      value = value.replace(regexNotNumberOrDot, '');
-      value = value.replace(regexLeadingNumber, '');
-    }
-    valueUpdateCB(label, value);
-  };
+			value = value.replace(regexNotNumberOrDot, '');
+			value = value.replace(regexLeadingNumber, '');
+		}
+		valueUpdateCB(label, value);
+	};
 
-  return (
-    <View style={inputStyle.inputContainer}>
-      <Text style={inputStyle.textContainer}>{label}</Text>
-      <NumberTextInput
-        label={label}
-        value={value}
-        valueUpdateCB={removeNonNumericValues}
-      />
-    </View>
-  );
+	return (
+		<View style={inputStyle.inputContainer}>
+			<Text style={inputStyle.textContainer}>{label}</Text>
+			<NumberTextInput
+				label={label}
+				value={value}
+				valueUpdateCB={removeNonNumericValues}
+			/>
+		</View>
+	);
 };
 
 FormLine.propTypes = {
-  label: PropTypes.string.isRequired,
-  value: PropTypes.number.isRequired,
-  valueUpdateCB: PropTypes.func.isRequired,
+	label: PropTypes.string.isRequired,
+	value: PropTypes.number.isRequired,
+	valueUpdateCB: PropTypes.func.isRequired
 };
 
 export default FormLine;
