+++ app/components/TotalContainer.jsx
@@ -1,28 +1,30 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { View, Text } from 'react-native';
-import { totalStyle } from '../Styles/style';
+import { totalStyle } from '../styles/style';
 
 const TotalContainer = ({ label, value }) => {
-  value = parseFloat(value);
-  function convertToDollarString(value) {
-    const dollar = `$ ${value.toFixed(2).toString()}`;
-    return dollar;
-  }
+	value = parseFloat(value);
+	function convertToDollarString(value) {
+		const dollar = `$ ${value.toFixed(2).toString()}`;
+		return dollar;
+	}
 
-  return (
-    <View style={totalStyle.wrapper}>
-      <View style={totalStyle.container}>
-        <Text style={totalStyle.textLabel}>{label}</Text>
-        <Text style={totalStyle.numFont}>{convertToDollarString(value)}</Text>
-      </View>
-    </View>
-  );
+	return (
+		<View style={totalStyle.wrapper}>
+			<View style={totalStyle.container}>
+				<Text style={totalStyle.textLabel}>{label}</Text>
+				<Text style={totalStyle.numFont}>
+					{convertToDollarString(value)}
+				</Text>
+			</View>
+		</View>
+	);
 };
 
 TotalContainer.propTypes = {
-  label: PropTypes.string.isRequired,
-  value: PropTypes.number.isRequired,
+	label: PropTypes.string.isRequired,
+	value: PropTypes.number.isRequired
 };
 
 export default TotalContainer;
