+++ app/index.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import MortgageFormContainer from './containers/MortgageFormContainer';
 
-const Application = () => (
-  <MortgageFormContainer />
-);
+const Application = () => <MortgageFormContainer />;
 
 export default Application;
