+++ package.json
@@ -24,10 +24,13 @@
     "babel-preset-expo": "~8.0.0",
     "eslint": "^6.7.2",
     "eslint-config-airbnb": "^18.0.1",
+    "eslint-config-prettier": "^6.7.0",
     "eslint-plugin-import": "^2.19.1",
     "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-react": "^7.17.0",
     "eslint-plugin-react-hooks": "^1.7.0",
+    "prettier": "^1.19.1",
     "redux-devtools": "^3.5.0"
   },
   "private": true
